Rename gui to observer
Previously we had CreateProjectObserver implement Gui.
Switching Gui to Observer meant that we had a naming overload
where a class had the same name as the interface it was trying
to implement. (CreateProjectObserver : CreateProjectObserver).

I took a stab at namespacing the interface under the create_project
function in the similar manner as this post suggest:
https://www.novoda.com/blog/better-class-naming/